Add some small test functions gathering statistics
They are for now displayed in the page action but should be displayed
along with the options instead.

The display to be useful should also have a percentage display of
secure / httpsOnly as it is the stat that need to be tracker.

Once the filtering is implemented it may be needed to have the stats
without the data that will disapear in a short time.